requirements.txt
torch==2.3.0
pillow==10.3.0
numpy==1.26.4